Definition of 'SpaTalk' class

An S4 class containing the data, meta, and results of inferred cell type compositions, LR pairs, and pathways.

Slots

data

A list containing the raw and normalized data.

meta

A list containing the raw and new meta data.

para

A list containing the parameters.

coef

A matrix containing the results of deconvolution.

cellpair

A list containing the cell-cell pairs based on the spatial distance.

dist

A matrix containing the Euclidean distance among cells.

lrpair

A data frame containing the inferred LR pairs.

tf

A data frame containing the TFs of receptors.

lr_path

A list containing the lrpairs and pathways.