Decomposing cell-cell communications for spatial transciptomics data

Identify the all cell-cell communications for single-cell or spot-based spatial transciptomics data with proximal ligand-receptor-target interactions.

Usage

dec_cci_all(
  object,
  n_neighbor = 10,
  min_pairs = 5,
  min_pairs_ratio = 0,
  per_num = 1000,
  pvalue = 0.05,
  co_exp_ratio = 0.1,
  if_doParallel = T,
  use_n_cores = NULL
)

Arguments

object

SpaTalk object after find_lr_path.

n_neighbor

Number of neighbor cells to select as the proximal cell-cell pair. Default is 10.

min_pairs

Min proximal cell-cell pairs between for sending and receiving cell types. Default is 5.

min_pairs_ratio

Min proximal cell-cell pairs ratio between for sending and receiving cell types. Default is 0.

per_num

Number of repeat times for permutation test. Default is 1000.

pvalue

Include the significantly proximal LR pairs with this cutoff of p value from permutation test. Default is 0.05.

co_exp_ratio

Min cell ratio in receiving cells with co-expressed source and target genes for predicting the downstream pathway activity.

if_doParallel

Use doParallel. Default is TRUE.

use_n_cores

Number of CPU cores to use. Default is all cores - 2.

Value

SpaTalk object containing the inferred LR pairs and pathways.