Find lrpairs and pathways

Find lrpairs and pathways with receptors having downstream targets and transcriptional factors.

Usage

find_lr_path(
  object,
  lrpairs,
  pathways,
  max_hop = NULL,
  if_doParallel = T,
  use_n_cores = NULL
)

Arguments

object

SpaTalk object generated from dec_celltype.

lrpairs

A data.frame of the system data containing ligand-receptor pairs of 'Human' and 'Mouse' from CellTalkDB.

pathways

A data.frame of the system data containing gene-gene interactions and pathways from KEGG and Reactome as well as the information of transcriptional factors.

max_hop

Max hop from the receptor to the downstream target transcriptional factor to find for receiving cells. Default is 3 for human and 4 for mouse.

if_doParallel

Use doParallel. Default is TRUE.

use_n_cores

Number of CPU cores to use. Default is all cores - 2.

Value

SpaTalk object containing the filtered lrpairs and pathways.