Get LR and downstream pathways

Get LR and downstream pathways and get p value of receptor-related pathways with LR-target genes by the Fisher-exact test.

Usage

get_lr_path(
  object,
  celltype_sender,
  celltype_receiver,
  ligand,
  receptor,
  min_gene_num = 5
)

Arguments

object

SpaTalk object generated from dec_cci.

celltype_sender

Name of celltype_sender.

celltype_receiver

Name of celltype_receiver.

ligand

Name of ligand from celltype_sender.

receptor

Name of receptor from celltype_receiver.

min_gene_num

Min genes number for each pathway.

Value

A list containing two data.frame. One is LR and downstream pathways, another is the p value of receptor-related pathways with LR-target genes.