Plot spatial distribution of a single cell type percent

Plot spatial distribution of a single predicted cell type percent for transcriptomics data

Usage

plot_st_celltype_percent(
  object,
  celltype,
  size = 1,
  color_low = NULL,
  color_mid = NULL,
  color_high = NULL,
  color_midpoint = NULL
)

Arguments

object

SpaTalk object generated from dec_celltype.

celltype

Name of cell type in the sc_celltype.

size

Point size. Default is 1.

color_low

Color for the lowest value.

color_mid

Color for the middle value for using scale_color_gradient2. Default is NULL.

color_high

Color for the highest value.

color_midpoint

Value for the middle scale. Default is NULL.