Pre-processing step: revising gene symbols

Revise genes according to NCBI Gene symbols updated in June 30, 2021 for count matrix, user-custom lrpairs data.frame, and user-custom pathways data.frame.

Usage

rev_gene(data = NULL, data_type = NULL, species = NULL, geneinfo = NULL)

Arguments

data

A data.frame or matrix or dgCMatrix containing count data each column representing a spot or a cell, each row representing a gene; Or a data.frame containing ligand-receptor pairs; Or a data.frame containing gene-gene interactions and pathways from KEGG and Reactome as well as the information of transcriptional factors.

data_type

A character to define the type of data, select 'count' for the data matrix, 'lrpairs' for the data.frame containing lrpairs, 'pathways' for the data.frame containing pathways.

species

Species of the data.'Human' or 'Mouse'.

geneinfo

A data.frame of the system data containing gene symbols of 'Human' and 'Mouse' updated on June 30, 2021 for revising count matrix.

Value

A new matrix or data.frame.